+++ tests/test_display_results.py
@@ -30,11 +30,11 @@ import pytest
 
 from servicex.servicex_client import (
     GuardList,
-    _display_results,
     _get_progress_options,
     ProgressBarFormat,
     ServiceXClient,
 )
+from servicex.app.main import _display_results
 
 
 def test_display_results_basic():
@@ -50,7 +50,7 @@ def test_display_results_basic():
         mock_console = Mock()
         mock_get_console.return_value = mock_console
 
-        with patch("servicex.servicex_client.Table"):
+        with patch("rich.table.Table"):
             _display_results(out_dict)
 
             # Just verify it was called - don't over-test internal details
@@ -90,3 +90,40 @@ def test_guardlist_basics():
     assert not error_list.valid()
     with pytest.raises(ReturnValueException):
         _ = error_list[0]
+
+
+def test_display_results_with_many_files():
+    from servicex.servicex_client import GuardList
+    from unittest.mock import patch, MagicMock
+
+    # Mock GuardList with more than 3 files to trigger lines 275-276
+    mock_guard_list = MagicMock(spec=GuardList)
+    mock_guard_list.valid.return_value = True
+    mock_guard_list.__iter__.return_value = iter(
+        [
+            "file1.parquet",
+            "file2.parquet",
+            "file3.parquet",
+            "file4.parquet",
+            "file5.parquet",
+        ]
+    )
+
+    out_dict = {"sample1": mock_guard_list}
+
+    with patch("rich.get_console") as mock_get_console:
+        mock_console = MagicMock()
+        mock_get_console.return_value = mock_console
+
+        with patch("rich.table.Table") as mock_table:
+            mock_table_instance = MagicMock()
+            mock_table.return_value = mock_table_instance
+
+            _display_results(out_dict)
+
+            # Verify that add_row was called with the truncated file list
+            mock_table_instance.add_row.assert_called_once()
+            call_args = mock_table_instance.add_row.call_args[0]
+            assert call_args[0] == "sample1"
+            assert call_args[1] == "5"
+            assert "... and 3 more files" in call_args[2]
